2023 days 11-13

+++ utils/utils.go
@@ -20,7 +20,7 @@ func Min(number1, number2 int) int {
 	return number2
 }
 
-func Abs(value int) int {
+func Abs(value int64) int64 {
 	if value >= 0 {
 		return value
 	}
